Abandon basket variables inclusion file
Abandon basket do not need all the data calls made in the normal dialogue global variables. But it need a call to SEGMENTATION_1_COL, which does not exist in the normal dialogue global variables file. So I have made a special inclusion file only to be used for abandon basket emails.